feat: makes write method concurrency safe

+++ src/core/disk.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"os"
 	"regexp"
+	"sync"
 )
 
 type Disk interface {
@@ -43,7 +44,18 @@ func (dc *DiskConfiguration) CreateDirectoryChain(chainPath string) (bool, error
 	return true, nil
 }
 
+// Global lock to ensure that only write operation runs at a time
+var diskWriteLock sync.Mutex
+
+// This method provides isolation to all the callers. It uses RWMutex
+// to prevent data overwrite or inconsistent data
 func (dc *DiskConfiguration) Write(fileName string, b []byte) error {
+	diskWriteLock.Lock()
+	defer diskWriteLock.Unlock()
+
+	if dc.Exists(fileName) {
+		return errors.New(RecordAlreadyExists)
+	}
 	err := os.WriteFile(fileName, b, DefaultFilePermission)
 	return err
 }
